♻️ refactor: update build:favicons script to use new faviconData.json location
The build:favicons script has been updated to use the new location of faviconData.json, which is now located in packages/assets-favicons/assets. This change was necessary to improve the build process and ensure that the correct file is used. Additionally, the .gitignore file has been updated to include the new location of the faviconData.json file.